modify status code and unit test

+++ WebApiBackend/WebApiBackend/Controllers/FlatController.cs
@@ -37,17 +37,7 @@ namespace WebApiBackend.Controllers
             _context = context;
         }
 
-        [Authorize]
-        [HttpGet("display")]
-        public ActionResult<List<DisplayMemberDTO>> GetMembers()
-        {
-            var identity = (ClaimsIdentity)HttpContext.User.Identity;
-            int userID = Int16.Parse(identity.FindFirst(ClaimTypes.NameIdentifier).Value);
-            var user = _context.User.Find(userID);
-            Flat flat = _context.Flat.Where(f => f.Id == user.FlatId).FirstOrDefault();
-            IQueryable members = _context.Entry(flat).Collection(f => f.Users).Query().OrderBy(u => u.FirstName);
-            return _MemberMapper.Map<List<DisplayMemberDTO>>(members);
-        }
+     
 
         /// <summary>
         /// GET Method - Get a list of members in the current users flat
@@ -107,36 +97,49 @@ namespace WebApiBackend.Controllers
         /// <summary>
         /// Delete Method -Remove a member out of a flat
         /// </summary>
-        /// <response code="201">Able to find the user and create a new flat for them</response>
-        /// <response code="401">Not an authorised user</response>
-        /// /// <response code="403">Attempting to create a new flat while already part of one</response>
-        /// <returns>The JSON representation of the users newly created flat</returns>
-        //Delete:api/delete/{deleteUsername}
+        /// <param name="deleteUserName">The username of a member being removed from the flat</param>
+        /// <response code="202">The user have successfully remove her/himself from the flat, and be redirected to create flat page</response>
+        /// <returns>The response body is the flat members </returns>
+
         [Authorize]
         [HttpDelete("{deleteUserName}")]
-        public ActionResult RemoveMember(string deleteUserName)
+        public ActionResult<FlatDTO> RemoveMember(string deleteUserName)
         {
             ClaimsIdentity identity = (ClaimsIdentity)HttpContext.User.Identity;
             int userId = Int32.Parse(identity.FindFirst(ClaimTypes.NameIdentifier).Value);
+           
             User user = _context.User.Find(userId);
-            User deleteUser = _context.User.Where(u => u.UserName == deleteUserName).First();
-            Flat flat = _context.Flat.Include(f => f.Users).First(f => f.Id == user.FlatId);
- 
-            var redirectToActionResult = new RedirectToActionResult("createFlat", "Flat", null);
+            User deleteUser = _context.User.FirstOrDefault(u => u.UserName == deleteUserName);
+            Flat flat = _context.Flat.Include(f => f.Users).FirstOrDefault(f => f.Id == user.FlatId);
+
+           
+            if (deleteUser == null || !flat.Users.Contains(deleteUser))
+            {
+
+                return NotFound("OOPS, user not in the flat.");
+            }
+
             flat.Users.Remove(deleteUser);
             _context.SaveChanges();
      
+            //code 202 retrun when the user is removing her/himself out from the flat
             if (deleteUser.Id == userId)
             {
+                //flat will be remove from system once empty 
                 if (flat.Users.Count == 0)
                 {
                     _context.Flat.Remove(flat);
                     _context.SaveChanges();
+                    
                 }
-                return redirectToActionResult;
+                Response.StatusCode = 202;
+                return null;
             }
-
-            return Ok(deleteUser.UserName + " has been removed from the flat.");
+            //return the new member list of the flat
+            return new FlatDTO(flat);
         }
+
+
+
     }
 }
